+++ CHANGES.md
@@ -1,4 +1,5 @@
 ## Version 2.8.12 - TBD
+  * Add color-blind friendly spectrum option for spectrogram (#275)
   * Fix start time precision in Seisan file export (#280)
   * Fix wrong year in wave panel for Dec 31
   * Handle error when map view selected for station with no lat/lon in metadata